Separate store files
