+++ pkg/dss/cockroach/store.go
@@ -4,16 +4,12 @@ import (
 	"context"
 	"database/sql"
 	"database/sql/driver"
-	"errors"
 	"fmt"
 	"strconv"
 	"time"
 
-	"github.com/golang/geo/s2"
-	"github.com/golang/protobuf/ptypes"
-	"github.com/lib/pq" // Pull in the postgres database driver
+	"github.com/golang/protobuf/ptypes" // Pull in the postgres database driver
 	dspb "github.com/steeling/InterUSS-Platform/pkg/dssproto"
-	"go.uber.org/multierr"
 )
 
 type scanner interface {
@@ -129,61 +125,6 @@ func (nt nullTime) Value() (driver.Value, error) {
 	return nt.Time, nil
 }
 
-type subscriptionsRow struct {
-	id                string
-	owner             string
-	url               string
-	typesFilter       sql.NullString
-	notificationIndex int32
-	lastUsedAt        pq.NullTime
-	beginsAt          pq.NullTime
-	expiresAt         pq.NullTime
-	updatedAt         time.Time
-}
-
-func (sr *subscriptionsRow) scan(scanner scanner) error {
-	return scanner.Scan(&sr.id,
-		&sr.owner,
-		&sr.url,
-		&sr.typesFilter,
-		&sr.notificationIndex,
-		&sr.lastUsedAt,
-		&sr.beginsAt,
-		&sr.expiresAt,
-		&sr.updatedAt,
-	)
-}
-
-func (sr *subscriptionsRow) toProtobuf() (*dspb.Subscription, error) {
-	result := &dspb.Subscription{
-		Id:    sr.id,
-		Owner: sr.owner,
-		Callbacks: &dspb.SubscriptionCallbacks{
-			IdentificationServiceAreaUrl: sr.url,
-		},
-		NotificationIndex: int32(sr.notificationIndex),
-		Version:           timestampToVersionString(sr.updatedAt),
-	}
-
-	if sr.beginsAt.Valid {
-		ts, err := ptypes.TimestampProto(sr.beginsAt.Time)
-		if err != nil {
-			return nil, err
-		}
-		result.Begins = ts
-	}
-
-	if sr.expiresAt.Valid {
-		ts, err := ptypes.TimestampProto(sr.expiresAt.Time)
-		if err != nil {
-			return nil, err
-		}
-		result.Expires = ts
-	}
-
-	return result, nil
-}
-
 func versionStringToTimestamp(s string) (time.Time, error) {
 	var t time.Time
 	nanos, err := strconv.ParseUint(s, versionBase, 64)
@@ -201,38 +142,6 @@ func (sr *subscriptionsRow) versionOK(version string) bool {
 	return version == "" || version == timestampToVersionString(sr.updatedAt)
 }
 
-// from or apply
-func (sr *subscriptionsRow) applyProtobuf(subscription *dspb.Subscription) error {
-	if subscription.Id != "" {
-		sr.id = subscription.Id
-	}
-
-	if subscription.Owner != "" {
-		sr.owner = subscription.Owner
-	}
-	if subscription.Callbacks.GetIdentificationServiceAreaUrl() != "" {
-		sr.url = subscription.Callbacks.GetIdentificationServiceAreaUrl()
-	}
-	if ts := subscription.GetBegins(); ts != nil {
-		begins, err := ptypes.Timestamp(ts)
-		if err != nil {
-			return err
-		}
-		sr.beginsAt.Time = begins
-		sr.beginsAt.Valid = true
-	}
-
-	if ts := subscription.GetExpires(); ts != nil {
-		expires, err := ptypes.Timestamp(ts)
-		if err != nil {
-			return err
-		}
-		sr.expiresAt.Time = expires
-		sr.expiresAt.Valid = true
-	}
-	return nil
-}
-
 // Store is an implementation of dss.Store using
 // Cockroach DB as its backend store.
 type Store struct {
@@ -242,579 +151,6 @@ type Store struct {
 // Close closes the underlying DB connection.
 func (s *Store) Close() error {
 	return s.DB.Close()
-}
-
-// insertSubscription inserts subscription into the store and returns
-// the resulting subscription including its ID.
-func (s *Store) insertSubscription(ctx context.Context, subscription *dspb.Subscription, cells s2.CellUnion) (*dspb.Subscription, error) {
-	const (
-		insertQuery = `
-		INSERT INTO
-			subscriptions
-		VALUES
-			($1, $2, $3, $4, $5, $6, $7, $8, transaction_timestamp())
-		RETURNING
-			*`
-		subscriptionCellQuery = `
-		INSERT INTO
-			cells_subscriptions
-		VALUES
-			($1, $2, $3, transaction_timestamp())
-		`
-	)
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	sr := &subscriptionsRow{}
-
-	sr.applyProtobuf(subscription)
-
-	if err := sr.scan(tx.QueryRowContext(
-		ctx,
-		insertQuery,
-		sr.id,
-		sr.owner,
-		sr.url,
-		sr.typesFilter,
-		sr.notificationIndex,
-		sr.lastUsedAt,
-		sr.beginsAt,
-		sr.expiresAt,
-	)); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	for _, cell := range cells {
-		if _, err := tx.ExecContext(ctx, subscriptionCellQuery, cell, cell.Level(), sr.id); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-	}
-
-	result, err := sr.toProtobuf()
-	if err != nil {
-		return nil, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-// updatesSubscription updates the subscription  and returns
-// the resulting subscription including its ID.
-func (s *Store) updateSubscription(ctx context.Context, subscription *dspb.Subscription, cells s2.CellUnion) (*dspb.Subscription, error) {
-	const (
-		// We use an upsert so we don't have to specify the column
-		updateQuery = `
-		UPSERT INTO
-		  subscriptions
-		VALUES
-			($1, $2, $3, $4, $5, $6, $7, $8, transaction_timestamp())
-		RETURNING
-			*`
-		subscriptionCellQuery = `
-		UPSERT INTO
-			cells_subscriptions
-		VALUES
-			($1, $2, $3, transaction_timestamp())
-		`
-		getQuery = `
-		SELECT * FROM
-			subscriptions
-		WHERE
-			id = $1`
-	)
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	sr := &subscriptionsRow{}
-
-	err = sr.scan(tx.QueryRowContext(ctx, getQuery, subscription.Id))
-
-	switch {
-	case err == sql.ErrNoRows: // Do nothing here.
-		return nil, multierr.Combine(err, tx.Rollback())
-	case err != nil:
-		return nil, multierr.Combine(err, tx.Rollback())
-	case !sr.versionOK(subscription.Version):
-		err := fmt.Errorf("version mismatch for subscription %s", subscription.Id)
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	sr.applyProtobuf(subscription)
-
-	if err := sr.scan(tx.QueryRowContext(
-		ctx,
-		updateQuery,
-		sr.id,
-		sr.owner,
-		sr.url,
-		sr.typesFilter,
-		sr.notificationIndex,
-		sr.lastUsedAt,
-		sr.beginsAt,
-		sr.expiresAt,
-	)); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	// TODO(steeling) we also need to delete any leftover cells.
-	for _, cell := range cells {
-		if _, err := tx.ExecContext(ctx, subscriptionCellQuery, cell, cell.Level(), sr.id); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-	}
-
-	result, err := sr.toProtobuf()
-	if err != nil {
-		return nil, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-// GetSubscription returns the subscription identified by "id".
-func (s *Store) GetSubscription(ctx context.Context, id string) (*dspb.Subscription, error) {
-	const (
-		subscriptionQuery = `
-		SELECT * FROM
-			subscriptions
-		WHERE
-			id = $1`
-	)
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	sr := &subscriptionsRow{}
-
-	if err := sr.scan(tx.QueryRowContext(ctx, subscriptionQuery, id)); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	result, err := sr.toProtobuf()
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-// DeleteSubscription deletes the subscription identified by "id" and
-// returns the deleted subscription.
-func (s *Store) DeleteSubscription(ctx context.Context, id, version string) (*dspb.Subscription, error) {
-	const (
-		blindQuery = `
-		DELETE FROM
-			subscriptions
-		WHERE
-			id = $1
-		RETURNING
-			*`
-
-		idempotentQuery = `
-		DELETE FROM
-			subscriptions
-		WHERE
-			id = $1
-			AND updated_at = $2
-		RETURNING
-			*`
-	)
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	sr := &subscriptionsRow{}
-	switch version {
-	case "":
-		if err := sr.scan(tx.QueryRowContext(ctx, blindQuery, id)); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-	default:
-		updatedAt, err := versionStringToTimestamp(version)
-		if err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-		if err := sr.scan(tx.QueryRowContext(ctx, idempotentQuery, id, updatedAt)); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-	}
-
-	result, err := sr.toProtobuf()
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-// SearchSubscriptions returns all subscriptions in "cells".
-func (s *Store) SearchSubscriptions(ctx context.Context, cells s2.CellUnion, owner string) ([]*dspb.Subscription, error) {
-	const (
-		subscriptionsInCellsQuery = `
-			SELECT
-				subscriptions.*
-			FROM
-				subscriptions
-			LEFT JOIN 
-				(SELECT DISTINCT cells_subscriptions.subscription_id FROM cells_subscriptions WHERE cells_subscriptions.cell_id = ANY($1))
-			AS
-				unique_subscription_ids
-			ON
-				subscriptions.id = unique_subscription_ids.subscription_id
-			WHERE
-				subscriptions.owner = $2`
-	)
-
-	if len(cells) == 0 {
-		return nil, errors.New("missing cell IDs for query")
-	}
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	rows, err := tx.QueryContext(ctx, subscriptionsInCellsQuery, pq.Array(cells), owner)
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-	defer rows.Close()
-
-	var (
-		row    = &subscriptionsRow{}
-		result = []*dspb.Subscription{}
-	)
-
-	for rows.Next() {
-		if err := row.scan(rows); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-		pb, err := row.toProtobuf()
-		if err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-
-		result = append(result, pb)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-func (s *Store) insertIdentificationServiceAreaUnchecked(ctx context.Context, serviceArea *dspb.IdentificationServiceArea, cells s2.CellUnion) (*dspb.IdentificationServiceArea, error) {
-	const (
-		subscriptionQuery = `
-		INSERT INTO
-			identification_service_areas
-		VALUES
-			($1, $2, $3, $4, $5, transaction_timestamp())
-		RETURNING
-			*`
-		subscriptionCellQuery = `
-		INSERT INTO
-			cells_identification_service_areas
-		VALUES
-			($1, $2, $3, transaction_timestamp())
-		`
-	)
-
-	isar := identificationServiceAreaRow{
-		id:    serviceArea.GetId(),
-		owner: serviceArea.GetOwner(),
-		url:   serviceArea.GetFlightsUrl(),
-	}
-
-	starts, err := ptypes.Timestamp(serviceArea.GetExtents().GetTimeStart())
-	if err != nil {
-		return nil, err
-	}
-	isar.startsAt = starts
-
-	ends, err := ptypes.Timestamp(serviceArea.GetExtents().GetTimeEnd())
-	if err != nil {
-		return nil, err
-	}
-	isar.endsAt = ends
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	if err := isar.scan(tx.QueryRowContext(
-		ctx,
-		subscriptionQuery,
-		isar.id,
-		isar.owner,
-		isar.url,
-		isar.startsAt,
-		isar.endsAt,
-	)); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	for _, cell := range cells {
-		if _, err := tx.ExecContext(ctx, subscriptionCellQuery, cell, cell.Level(), isar.id); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-	}
-
-	result := &dspb.IdentificationServiceArea{
-		Id:         isar.id,
-		Owner:      isar.owner,
-		FlightsUrl: isar.url,
-	}
-
-	ts, err := ptypes.TimestampProto(isar.startsAt)
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-	result.Extents = &dspb.Volume4D{
-		SpatialVolume: serviceArea.GetExtents().GetSpatialVolume(),
-		TimeStart:     ts,
-	}
-
-	ts, err = ptypes.TimestampProto(isar.endsAt)
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-	result.Extents.TimeEnd = ts
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return result, err
-}
-
-// DeleteIdentificationServiceArea deletes the IdentificationServiceArea identified by "id" and owned by "owner".
-// Returns the delete IdentificationServiceArea and all Subscriptions affected by the delete.
-func (s *Store) DeleteIdentificationServiceArea(ctx context.Context, id string, owner string) (*dspb.IdentificationServiceArea, []*dspb.SubscriberToNotify, error) {
-	const (
-		getAffectedCellsAndSubscriptions = `
-			SELECT
-				cells_identification_service_areas.cell_id,
-				affected_subscriptions.subscription_id
-			FROM
-				cells_identification_service_areas
-			LEFT JOIN
-				(SELECT DISTINCT cell_id, subscription_id FROM cells_subscriptions)
-			AS
-				affected_subscriptions
-			ON
-				affected_subscriptions.cell_id = cells_identification_service_areas.cell_id
-			WHERE
-				cells_identification_service_areas.identification_service_area_id = $1
-		`
-		getSubscriptionDetailsForAffectedCells = `
-			SELECT
-				id, url, notification_index
-			FROM
-				subscriptions
-			WHERE
-				id = ANY($1)
-			AND
-				owner != $2
-			AND
-				begins_at IS NULL OR transaction_timestamp() >= begins_at
-			AND
-				expires_at IS NULL OR transaction_timestamp() <= expires_at
-		`
-		deleteIdentificationServiceAreaQuery = `
-			DELETE FROM
-				identification_service_areas
-			WHERE
-				id = $1
-			AND
-				owner = $2
-			RETURNING
-				*
-		`
-	)
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var (
-		cells         []int64
-		subscriptions []string
-		cell          int64
-		subscription  string
-	)
-
-	rows, err := tx.QueryContext(ctx, getAffectedCellsAndSubscriptions, id)
-	if err != nil {
-		return nil, nil, multierr.Combine(err, tx.Rollback())
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		if err := rows.Scan(&cell, &subscription); err != nil {
-			return nil, nil, multierr.Combine(err, tx.Rollback())
-		}
-		cells = append(cells, cell)
-		subscriptions = append(subscriptions, subscription)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	isar := &identificationServiceAreaRow{}
-	if err := isar.scan(tx.QueryRowContext(ctx, deleteIdentificationServiceAreaQuery, id, owner)); err != nil {
-		// This error condition will be triggered if the owner does not match.
-		multierr.Combine(err, tx.Rollback())
-	}
-
-	var (
-		subscribers []subscriberRow
-		subscriber  subscriberRow
-	)
-
-	rows, err = tx.QueryContext(ctx, getSubscriptionDetailsForAffectedCells, pq.Array(subscriptions), owner)
-	if err != nil {
-		return nil, nil, multierr.Combine(err, tx.Rollback())
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		if err := rows.Scan(&subscriber.id, &subscriber.url, &subscriber.notificationIndex); err != nil {
-			return nil, nil, multierr.Combine(err, tx.Rollback())
-		}
-
-		subscribers = append(subscribers, subscriber)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	isa, err := isar.toProtobuf()
-	if err != nil {
-		return nil, nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	subscribersToNotify := []*dspb.SubscriberToNotify{}
-	for _, subscriber := range subscribers {
-		subscriberToNotify, err := subscriber.toProtobuf()
-		if err != nil {
-			return nil, nil, multierr.Combine(err, tx.Rollback())
-		}
-		subscribersToNotify = append(subscribersToNotify, subscriberToNotify)
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, nil, err
-	}
-
-	return isa, subscribersToNotify, nil
-}
-
-// SearchIdentificationServiceAreas searches IdentificationServiceArea
-// instances that intersect with "cells" and, if set, the temporal volume
-// defined by "earliest" and "latest".
-func (s *Store) SearchIdentificationServiceAreas(ctx context.Context, cells s2.CellUnion, earliest *time.Time, latest *time.Time) ([]*dspb.IdentificationServiceArea, error) {
-	const (
-		serviceAreasInCellsQuery = `
-			SELECT
-				identification_service_areas.*
-			FROM
-				identification_service_areas
-			JOIN 
-				(SELECT DISTINCT
-					cells_identification_service_areas.identification_service_area_id
-				FROM
-					cells_identification_service_areas
-				WHERE
-					cells_identification_service_areas.cell_id = ANY($1)
-				)
-			AS
-				unique_identification_service_areas
-			ON
-				identification_service_areas.id = unique_identification_service_areas.identification_service_area_id
-			WHERE
-				COALESCE(identification_service_areas.starts_at >= $2, true)
-			AND
-				COALESCE(identification_service_areas.ends_at <= $3, true)`
-	)
-
-	if len(cells) == 0 {
-		return nil, errors.New("missing cell IDs for query")
-	}
-
-	cids := make([]int64, len(cells))
-	for i, cid := range cells {
-		cids[i] = int64(cid)
-	}
-
-	tx, err := s.Begin()
-	if err != nil {
-		return nil, err
-	}
-
-	rows, err := tx.QueryContext(ctx, serviceAreasInCellsQuery, pq.Int64Array(cids), earliest, latest)
-	if err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-	defer rows.Close()
-
-	var (
-		result = []*dspb.IdentificationServiceArea{}
-		isar   = &identificationServiceAreaRow{}
-	)
-
-	for rows.Next() {
-		if err := isar.scan(rows); err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-
-		pb, err := isar.toProtobuf()
-		if err != nil {
-			return nil, multierr.Combine(err, tx.Rollback())
-		}
-		result = append(result, pb)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, multierr.Combine(err, tx.Rollback())
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return result, nil
 }
 
 // Bootstrap bootstraps the underlying database with required tables.
